refactor: resolve test issues by changing imports
