+++ packages/common/e2e/catalog-collection.e2e.ts
@@ -1,5 +1,4 @@
-import { cloneObject, IHubCatalog, IHubCollection } from "../src";
-import { Catalog } from "../src/search";
+import { Catalog, IHubCatalog, IHubCollection } from "../src/search";
 import { Collection } from "../src/search/Collection";
 import Artifactory from "./helpers/Artifactory";
 import config from "./helpers/config";
