+++ packages/common/src/core/types/IHubProject.ts
@@ -1,10 +1,11 @@
-import { IHubTimeline, IHubItemEntity } from "./index";
 import {
   IWithLayout,
   IWithPermissions,
   IWithSlug,
   IWithCatalog,
 } from "../traits/index";
+import { IHubItemEntity } from "./IHubItemEntity";
+import { IHubTimeline } from "./IHubTimeline";
 
 /**
  * Defines the properties of a Hub Project object
